For AnagramDictionary: ----- For function getAnagramsWithOneMoreLetter: Use a for loop to iterate through the letters of the alphabet. For each letter, make a key containing the base word plus the letter. Make a key from the word + letter by sorting the letters.
Using that key, get a list of words from the HashMap lettersToWord that
shares the same characters. By default, this is already an anagram of
the word plus the letter.

Add each word from the list from lettersToWord into a new List, named
result.

The result List stores all of the anagrams of the word + letter, and
is returned after the letters of the alphabet is iterated through and
lettersTo Word is checked each time.